+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## version 0.2.0+ (Unreleased)
 - add --version option
 - split help document
+- add some subcommands on v2
+- add `insdeltasec` subcommand
 
 ## version 0.2.0 (2018-02-25)
 - add -v1, -v2 options